% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CutEdges.R
\name{CutEdges}
\alias{CutEdges}
\title{Remove all edges that lie between pairs of clusters}
\usage{
CutEdges(igraph_object, cluster_pairs = NULL, cluster.col = "cluster")
}
\arguments{
\item{igraph_object}{An igraph object with cluster results stored.}

\item{cluster_pairs}{A list containing two-element vectors indicating
pairs of cluster names where we want to remove edges between.}

\item{cluster.col}{A character value with the vertex attribute name in
the igraph object with clustering results.}
}
\value{
an updated igraph object with some edges removed
}
\description{
Remove all edges that lie between pairs of clusters
}
\examples{
# Create a coordinate data frame
df = rbind(expand.grid(1:5, 1:5), expand.grid(9:11, 9:11))
colnames(df) = c('x', 'y')

# Create an igraph object with these coordinates
ig = SpotGraph(df)

# Perform clustering and add those cluster assignments to the igraph object
cl = igraph::cluster_louvain(ig)$membership
ig = igraph::set_vertex_attr(ig, 'cluster', value = factor(cl))

# Examine graph before removing edges
# library(ggnetwork)
# ggplot(ig, aes(x=x, y=y, xend=xend, yend=yend)) +
#  geom_edges() +
#  geom_nodes(aes(color = cluster))

# Remove edges between clusters 1 and 2, and between clusters 3 and 4
ig = CutEdges(igraph_object = ig,
              cluster_pairs = list(c(1,2), c(1,3)),
              cluster.col = 'cluster')

# Examine graph after removing edges
# ggplot(ig, aes(x=x, y=y, xend=xend, yend=yend)) +
#   geom_edges() +
#   geom_nodes(aes(color = cluster))
}
