% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_threshold.R
\name{get_threshold}
\alias{get_threshold}
\title{Calculate a threshold given some set of statistics}
\usage{
get_threshold(
  stats,
  percentile1 = 0.05,
  percentile2 = 0.95,
  log = T,
  print.plot = T
)
}
\arguments{
\item{stats}{a numeric vector.}

\item{percentile1}{the percentile to set as the lower bound for calculating a local minimum.}

\item{percentile2}{the percentile to set as the upper bound for calculating a local minimum.}

\item{log}{logical (default \code{TRUE}), whether to use the log of the input statistics before
calculating a threshold; useful for calculating threshold for filtering spots based on
number of transcripts detected per-spot.}

\item{print.plot}{logical (default \code{TRUE}), whether to print a plot of approximated function
with intervals (red) and threshold (blue) plotted.}
}
\value{
The local minimum of the approximated function between the percentiles
set by \code{percentile1} and \code{percentile2}. A plot of the approximated function will
also be printed if \code{print.plot == TRUE} with the returned value and intervals plotted.
}
\description{
Calculate a threshold given some set of statistics
}
\examples{
\dontrun{
# To set a threshold for number of transcripts per spot to filter low quality spots:
seu_obj = scc_s1
nCount = seu_obj$nCount_Spatial
thres = get_threshold(nCount)
seu_obj_filtered = subset(seu_obj, nCount_Spatial > thres)
}
}
