% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeighborhoodCenters.R
\name{NeighborhoodCenters}
\alias{NeighborhoodCenters}
\title{Identify center of some neighborhood of spots, e.g., the center of a tumor}
\usage{
NeighborhoodCenters(igraph_object = NULL, is_neighborhood)
}
\arguments{
\item{igraph_object}{An igraph object}

\item{is_neighborhood}{A boolean vector (\code{TRUE}/\code{FALSE}) of equal length as the number of rows
in \code{coord}. Centers will be identified for each neighborhood of spots labeled \code{TRUE}.}
}
\value{
a list with the following elements:
\enumerate{
\item \code{eigen.scores}: a numeric vector with the output from igraph::centr_eigen, with a
score for each spot, where \code{max(eigen.scores)} within each neighborhood will
identify its center.
\item \code{centers}: a character vector with barcode ids of spots that were identified as the
center of each neighborhood of spots.
\item \code{boundaries}: a character vector with barcode ids of spots that were identified be on
the boundary of each neighborhood of spots.
}
}
\description{
Score spots using igraph::centr_eigen to identify spots in the center of some region/neighborhood.
}
\examples{
\dontrun{
df = rbind(expand.grid(1:5, 1:5), expand.grid(9:11, 9:11))
colnames(df) = c('x', 'y')
nspots = dim(df)[1]
rownames(df) = paste0('spot', 1:nspots)

igraph::V(ig)$cluster = igraph::cluster_leiden(ig, resolution = 0)$membership
is_cluster = igraph::V(ig)$cluster=='1'
names(is_cluster) = names(igraph::V(ig))

res = NeighborhoodCenters(coord = df, is_neighborhood = is_cluster)
}
}
