% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpotGraph.R
\name{SpotGraph}
\alias{SpotGraph}
\title{Creates an igraph object given x,y coordinates}
\usage{
SpotGraph(
  coord,
  dist.buffer = 1.05,
  max.dist = NULL,
  cluster = F,
  resolution = 0.5
)
}
\arguments{
\item{coord}{A two-column data.frame or matrix, where each column contains x or y coordinates.}

\item{dist.buffer}{Influences the maximum distance a node is allowed to be from another
node to be considered a neighbor. Only used if max.dist = NULL.}

\item{max.dist}{The furthest a node should be from another node to be considered neighbors.
If NULL (default), the shortest distance between any two nodes is the minimum distance (min.dist)
required for two nodes to be considered neighbors, and the max.dist is calculated as the hypotenuse
of a triangle with two sides of length min.dist, and then is multiplied by dist.buffer.}

\item{cluster}{Whether clustering should be performed on the igraph object.}

\item{resolution}{If cluster = T, a higher number will return more clusters and
a lower number will return fewer clusters. Clustering results are stored in the
returned igraph object under the name 'iglouvain_cluster' and can be accessed
by vertex_attr(igraph, 'iglouvain_cluster').}
}
\value{
an igraph object, where each vertex (i.e., node) corresponds to each
row in the coord input, with un-weighted edges between vertices that are
immediately adjacent to each other.
}
\description{
Create an igraph object with x,y coordinates, assuming all adjacent
spots are equidistant from each other, and drawing edges where spots
are immediately adjacent to each other.
}
\examples{
\dontrun{
# Create a coordinate data frame with two isolated groups of points
df = rbind(expand.grid(1:5, 1:5), expand.grid(9:11, 9:11))
colnames(df) = c('x', 'y')

# Preview the points on a grid with their corresponding x,y coordinates
ggplot(df, aes(x = x,y=y)) +
  geom_point()

# Identify points that are immediately adjacent to each other and create an igraph object
ig = SpotGraph(df)

# Optionally view the network with ggnetwork
# library(ggnetwork)
ggplot(ig, aes(x=x, y=y, xend=xend, yend=yend)) +
  geom_edges() +
  geom_nodes()
}
}
