% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpotGraphs-package.R
\docType{package}
\name{SpotGraphs-package}
\alias{SpotGraphs}
\alias{SpotGraphs-package}
\title{SpotGraphs: Generate igraph Objects With Visium Spot Coordinates For Basic Analysis Tasks}
\description{
Use x,y coordinates from 10X Visium data to create an igraph object, which can be useful to identify which spots are immediately neighboring each other and to provide a simple, lightweight framework to work with spatial communities of spots.
}
\author{
\strong{Maintainer}: Alex Lee \email{alee262@jhmi.edu} (\href{https://orcid.org/0000-0002-1825-2439}{ORCID})

Authors:
\itemize{
  \item David Sanin \email{david.sanin@jhu.edu} (\href{https://orcid.org/0000-0003-0188-7267}{ORCID})
}

}
\keyword{internal}
