% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanSlide.R
\name{CleanSlide}
\alias{CleanSlide}
\title{Identify spots isolated from the largest groups of immediately adjacent spots}
\usage{
CleanSlide(coord, nCount)
}
\arguments{
\item{coord}{A two-column data.frame or matrix, where each column contains x or y coordinates.}

\item{nCount}{A named vector of transcript counts per spot, where names are
spot barcodes/ids, and each element should match each row of coord.}
}
\value{
A data.frame with the following columns:
\enumerate{
\item ig_cluster: igraph modularity maximization results
\item cluster_nCount: summed transcript counts within each 'ig_cluster' group
\item threshold: whether the respective 'ig_cluster' group passed the
automatically determiend threshold
\item barcode: the spot id or barcode sequence
}
}
\description{
Identify spots in a 10X Visium Seurat object that are potentially low
quality. Mostly useful for slides where there are large areas without tissue
and standard spot-filtering based on nCounts is insufficient.
\enumerate{
\item Identify neighboring spots on x,y coordinates with SpotGraph()
\item Performing modularity maximization to identify clusters
\item Calculate total number of transcripts detected in each cluster (nCount)
\item Set a threshold to identify which clusters should be filtered out
}
}
\examples{
# Create a coordinate data frame with two isolated groups of points
df = rbind(expand.grid(1:5, 1:5), expand.grid(9:11, 9:11))
colnames(df) = c('x', 'y')
nspots = dim(df)[1]
rownames(df) = paste0('spot', 1:nspots)

# Create some vector of transcript counts
spotcounts = rbinom(25, 10000, 0.3)
spotcounts = c(spotcounts, rbinom(9, 100, 0.3))
names(spotcounts) = rownames(df)

# The resulting data.frame should indicate that the smaller
# cluster of 3x3 points did not pass the threshold, while the
# larger group of 5x5 points should have passed the threshold.
res = CleanSlide(coord = df, nCount = spotcounts)
}
