% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanSlide.R
\name{CleanSlide}
\alias{CleanSlide}
\title{Identify spots isolated from the largest groups of immediately adjacent spots}
\usage{
CleanSlide(obj)
}
\arguments{
\item{obj}{A Seurat object with 10X Visium data}
}
\value{
A Seurat object with several columns of meta-data added:
\enumerate{
\item igraph modularity maximization results stored in 'ig_cluster'
\item summed transcript counts within each 'ig_cluster' group in 'cluster_nCount'
\item whether the respective 'ig_cluster' group passed the automatic threshold
}
}
\description{
Identify spots in a 10X Visium Seurat object that are potentially low
quality. Mostly useful for slides where there are large areas without tissue
and standard spot-filtering based on nCounts is insufficient.
\enumerate{
\item Identify neighboring spots on x,y coordinates with SpotGraph()
\item Performing modularity maximization to identify clusters
\item Calculate total number of transcripts detected in each cluster (nCount)
\item Set a threshold to identify which clusters should be filtered out
}
}
\examples{
library(Seurat)
obj = CleanSlide(obj)
SpatialDimPlot(obj, group.by = 'ig_cluster')
}
