% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeighborhoodCenters.R
\name{NeighborhoodCenters}
\alias{NeighborhoodCenters}
\title{Identify center of some neighborhood of spots, e.g., the center of a tumor}
\usage{
NeighborhoodCenters(coord = NULL, is_neighborhood)
}
\arguments{
\item{coord}{A two-column data.frame or matrix, where each column contains x or y coordinates,
where the rownames are spot barcodes/ids.}

\item{is_neighborhood}{A boolean vector (\code{TRUE}/\code{FALSE}) of equal length as the number of rows
in \code{coord}. Centers will be identified for each neighborhood of spots labeled \code{TRUE}.}
}
\value{
a list with the following elements:
\enumerate{
\item \code{eigen.scores}: a numeric vector with the output from igraph::centr_eigen, with a
score for each spot, where \code{max(eigen.scores)} within each neighborhood will
identify its center.
\item \code{centers}: a character vector with barcode ids of spots that were identified as the
center of each neighborhood of spots.
\item \code{boundaries}: a character vector with barcode ids of spots that were identified be on
the boundary of each neighborhood of spots.
}
}
\description{
Identify center of some neighborhood of spots, e.g., the center of a tumor
}
\examples{
\dontrun{}
}
