% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialPlotGraph.R
\name{SpatialPlotGraph}
\alias{SpatialPlotGraph}
\title{Visualize igraph vertex attribute on tissue x,y coordinates}
\usage{
SpatialPlotGraph(
  igraph_object = NULL,
  coord = NULL,
  group.by = NULL,
  label = FALSE,
  linewidth = 0.5,
  pt.size = 1.6
)
}
\arguments{
\item{igraph_object}{An igraph object}

\item{coord}{(optional) A two-column data.frame or matrix, where each column contains x or y coordinates.
if not provided, will look for \code{coord_x} and \code{coord_y} attributes in the provided igraph object.}

\item{group.by}{Some vertex attribute to plot onto tissue coordinates. Must be
present in \code{vertex_attr(ig)}.}

\item{label}{Logical, whether to label groups on the plot based on \code{group.by}.}

\item{linewidth}{Numeric input to determine width of segments between nodes. Used in \verb{geom_segment().}}

\item{pt.size}{Numeric input to determine size of each point on the plot. Used as the size parameter in \code{geom_point()}.}
}
\value{
a ggplot object
}
\description{
Visualize igraph vertex attribute on tissue x,y coordinates
}
\examples{
\dontrun{
# Create a coordinate data frame and create and igraph object
df = rbind(expand.grid(1:5, 1:5), expand.grid(9:11, 9:11))
colnames(df) = c('x', 'y')
ig = SpotGraph(df)

# Plot cluster results
SpatialPlotGraph(igraph_object = ig, group.by = 'is_boundary')
}
}
