% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialPlotGraph.R
\name{SpatialPlotGraph}
\alias{SpatialPlotGraph}
\title{Visualize igraph vertex attribute on tissue x,y coordinates}
\usage{
SpatialPlotGraph(igraph_object, coord, group.by, label = T)
}
\arguments{
\item{igraph_object}{An igraph object}

\item{coord}{A two-column data.frame or matrix, where each column contains x or y coordinates.}

\item{group.by}{Some vertex attribute to plot onto tissue coordinates. Must be
present in vertex_attr(ig).}

\item{label}{Logical, whether to label groups on the plot based on group.by}
}
\value{
a ggplot object
}
\description{
Visualize igraph vertex attribute on tissue x,y coordinates
}
\examples{
# Create a coordinate data frame and create and igraph object
df = rbind(expand.grid(1:5, 1:5), expand.grid(9:11, 9:11))
colnames(df) = c('x', 'y')
ig = SpotGraph(df, cluster = T)

# Plot cluster results
SpatialPlotGraph(igraph_object = ig, coord = df, group.by = 'iglouvain_cluster')
}
